+++ src/core/se/jbee/inject/Injector.java
@@ -1,6 +1,6 @@
 /*
- *  Copyright (c) 2012-2019, Jan Bernitt 
- *			
+ *  Copyright (c) 2012-2019, Jan Bernitt
+ *	
  *  Licensed under the Apache License, Version 2.0, http://www.apache.org/licenses/LICENSE-2.0
  */
 package se.jbee.inject;
@@ -53,4 +53,5 @@ public interface Injector {
 	default <T> T resolve(Instance<T> inst) {
 		return resolve(dependency(inst));
 	}
+
 }