+++ src/core/se/jbee/inject/bind/BinderModule.java
@@ -37,6 +37,7 @@ public abstract class BinderModule extends InitializedBinder
 		if (basis != null)
 			bootstrap.install(basis);
 		bootstrap.install(DefaultScopes.class);
+		bootstrap.install(ExtensionModule.class);
 		bootstrap.install(this);
 	}