+++ src/core/se/jbee/inject/bind/BinderModuleWith.java
@@ -27,6 +27,7 @@ public abstract class BinderModuleWith<T> extends InitializedBinder
 	@Override
 	public final void bootstrap(Bootstrapper bootstrap) {
 		bootstrap.install(DefaultScopes.class);
+		bootstrap.install(ExtensionModule.class);
 		bootstrap.install(this);
 	}