+++ src/test/se/jbee/inject/bind/SuiteBind.java
@@ -25,7 +25,7 @@ import org.junit.runners.Suite.SuiteClasses;
 		TestEditionPackageBinds.class, TestPubSubBinds.class,
 		TestSetterInitialisationBinds.class, TestDynamicInitialiserBinds.class,
 		TestMacroBinds.class, TestGeneratorBinds.class,
-		TestDiskScopeBinds.class })
+		TestDiskScopeBinds.class, TestConfigBinds.class })
 public class SuiteBind {
 	// all tests in the bind package
 }